+++ Code/GINAC/ginac/power.cpp
@@ -983,11 +983,11 @@ ex power::expand_add(const add & a, long n, unsigned options)
 		// Multinomial expansion of power(+(x,...,z;0),k)*c^(n-k):
 		// Iterate over all partitions of k with exactly as many parts as
 		// there are symbolic terms in the basis (including zero parts).
-		partition_with_zero_parts_generator partitions(k, a.seq.size());
+		partition_with_zero_parts_generator partitions(k, static_cast<int>(a.seq.size()));
 		do {
 			const std::vector<unsigned>& partition = partitions.get();
 			// All monomials of this partition have the same number of terms and the same coefficient.
-			const unsigned msize = std::count_if(partition.begin(), partition.end(), [](int i) { return i > 0; });
+			const long long msize = std::count_if(partition.begin(), partition.end(), [](int i) { return i > 0; });
 			const numeric coeff = multinomial_coefficient(partition) * binomial_coefficient;
 
 			// Iterate over all compositions of the current partition.
