+++ Code/NestedSum/nestedsums/transcendental_C.cc
@@ -821,7 +821,7 @@ GINAC_BIND_UNARCHIVER(transcendental_sum_type_C);
     if ( is_a<mul>(subsum) || is_a<ncmul>(subsum) )
       {
 	int j;
-	int up_limit = subsum.nops();
+	int up_limit = static_cast<int>(subsum.nops());
 	exvector prefactor;
 	exvector hard_core;
 	prefactor.reserve(up_limit);
@@ -1020,7 +1020,7 @@ GINAC_BIND_UNARCHIVER(transcendental_sum_type_C);
     if ( is_a<mul>(letter) || is_a<ncmul>(letter) )
       {
 	int j;
-	int up_limit = letter.nops();
+	int up_limit = static_cast<int>(letter.nops());
 	exvector prefactor;
 	exvector hard_core;
 	prefactor.reserve(up_limit);
